Convert Durations Using Proper Converter
Implement the conversion of all duration-like fields using the
"TenThousandthSecondDurationConverter" so the values can be expressed as
"TimeSpan" instead of just a "long" value.
